# crypto.ecdsa: improves performances of `.public_key` call 