<!--

Please title your PR as follows: `module: description` (e.g. `time: fix date format`).
Always start with the thing you are fixing, then describe the fix.
Don't use past tense (e.g. "fixed foo bar").

Explain what your PR does and why.

If you are adding a new function, please document it and add tests:

```
// foo does foo and bar
fn foo() {

// file_test.v
fn test_foo() {
    assert foo() == ...
    ...
}
```

If you are fixing a bug, please add a test that covers it.

Before submitting a PR, please run `v test-all` .
See also `TESTS.md`.

I try to process PRs as soon as possible. They should be handled within 24 hours.

Applying labels to PRs is not needed.

Thanks a lot for your contribution!

-->

Even after this [PR](https://github.com/vlang/v/pull/23899), `PrivateKey.new` has been improved, overall of `generate_key` routine still has an overhead. In my test, its overhead happens on `.public_key` call. The overhead is about $\pm$ 310  µs (variesly).
After some digs to search the alternative approachs, i test 3 of them, for optimizing `public_key` part

- With the current approach, serializing using `d2i` based function
- Serializing and unserializing with `pem` based approach
- With duplicating key and clears out private info from the key.

With these approach, this is my test number

```bash
Average time using public_key_using_d2i (current)       : 288 µs
Average time using public_key_using_pem                 : 435 µs
Average time using public_key_using_dup_n_set_null      : 5 µs
```

The last approach looks like promising to be implemented and replacing the current approach. The benchmark test
show up the result

```bash
Benchmarking PrivateKey.public_key() (after patch)...
Average PrivateKey.public_key() (after patch) time: 5 µs
```

Its not a bad result, from $\pm$ 310  µs  goes down into  $\pm$  5 µs
This PR contains the third approach to replace the current of `public_key` method and the test still pass. Third approach was done by duplicating current private key and stripout private bits by replacing it with new empty bignum.

Thx